+++ src/aspnetcore/Kontent.Ai.AspNetCore/Webhooks/WebhookOptions.cs
@@ -3,7 +3,7 @@ namespace Kontent.Ai.AspNetCore.Webhooks;
 /// <summary>
 /// A configuration object that allows to adjust the Kontent.ai webhook behavior.
 /// </summary>
-public class WebhookOptions
+public sealed class WebhookOptions
 {
     /// <summary>
     /// Webhook secret used for webhook signature validation.